+++ R/geography.R
@@ -50,7 +50,7 @@ validate_coord <- function(unit_code, lat, lon) {
 
 #' Return UTM Zone
 #'
-#' @description `long2UTM()` take a longitude coordinate and returns the corresponding UTM zone.
+#' @description `convert_long_2_utm()` take a longitude coordinate and returns the corresponding UTM zone.
 #'
 #' @details Input a longitude (decimal degree) coordinate and this simple function returns the number of the UTM zone where that point falls.
 #'
@@ -58,7 +58,28 @@ validate_coord <- function(unit_code, lat, lon) {
 #'
 #' @return The function returns a numeric UTM zone (between 1 and 60).
 #' @export
+convert_long_to_utm <- function(lon) {
+  ## Function to get the UTM zone for a given longitude
+  return((floor((lon + 180) / 6) %% 60) + 1)
+}
+
+#' Return UTM Zone
+#'
+#' @description 
+#' `r lifecycle::badge("deprecated")`
+#' `long2UTM` was deprecated in favor of `convert_long_to_utm()` to enforce a consistent function naming pattern across the package and to conform to the tidyverse style guide.
+#' 
+#' `long2UTM()` take a longitude coordinate and returns the corresponding UTM zone.
+#'
+#' @details Input a longitude (decimal degree) coordinate and this simple function returns the number of the UTM zone where that point falls.
+#'
+#' @param lon - Decimal degree longitude value
+#'
+#' @return The function returns a numeric UTM zone (between 1 and 60).
+#' @keywords internal
+#' @export
 long2UTM <- function(lon) {
+  lifecycle::deprecate_warn("0.1.0.3", "long2UTM()", "convert_long_to_utm()")
   ## Function to get the UTM zone for a given longitude
   return((floor((lon + 180) / 6) %% 60) + 1)
 }
@@ -162,7 +183,7 @@ fuzz_location <- function(lat, lon, coord_ref_sys = 4326, fuzz_level = "Fuzzed -
 
 #' Coordinate Conversion from UTM to Latitude and Longitude
 #'
-#' @description utm_to_ll takes your dataframe with UTM coordinates in separate Easting and Northing columns, and adds on an additional two columns with the converted decimalLatitude and decimalLongitude coordinates using the reference coordinate system WGS84.
+#' @description `convert_utm_to_ll()` takes your dataframe with UTM coordinates in separate Easting and Northing columns, and adds on an additional two columns with the converted decimalLatitude and decimalLongitude coordinates using the reference coordinate system WGS84.
 #'
 #' @details Define the name of your dataframe, the easting and northing columns within it, the UTM zone within which those coordinates are located, and the reference coordinate system (datum). UTM Northing and Easting columns must be in separate columns prior to running the function. If a datum is not defined, the function will default to "WGS84". If there are missing coordinates in your dataframe they will be preserved, however they will be moved to the end of your dataframe. Note that some parameter names are not in snake_case but instead reflect DarwinCore naming conventions.
 #'
@@ -177,7 +198,7 @@ fuzz_location <- function(lat, lon, coord_ref_sys = 4326, fuzz_level = "Fuzzed -
 #'
 #' @examples
 #' \dontrun{
-#' utm_to_ll(
+#' convert_utm_to_ll(
 #'   df = mydataframe,
 #'   EastingCol = "EastingCoords",
 #'   NorthingCol = "NorthingCoords",
@@ -185,7 +206,7 @@ fuzz_location <- function(lat, lon, coord_ref_sys = 4326, fuzz_level = "Fuzzed -
 #'   datum = "WGS84"
 #' )
 #' }
-utm_to_ll <- function(df, EastingCol, NorthingCol, zone, datum = "WGS84") {
+convert_utm_to_ll <- function(df, EastingCol, NorthingCol, zone, datum = "WGS84") {
   Base <- as.data.frame(df)
   Base <- dplyr::rename(Base, "b" = EastingCol, "a" = NorthingCol)
   Mid <- Base[!is.na(Base$"b" & Base$"a"), ]
@@ -214,6 +235,73 @@ utm_to_ll <- function(df, EastingCol, NorthingCol, zone, datum = "WGS84") {
   df <- dplyr::rename(df,
     EastingCol = "b", NorthingCol = "a",
     "decimalLongitude" = x, "decimalLatitude" = y
+  )
+  return(df)
+}
+
+#' Coordinate Conversion from UTM to Latitude and Longitude
+#'
+#' @description 
+#' `r lifecycle::badge("deprecated")`
+#' `utm_to_ll()` was deprecated in favor of `convert_utm_to_ll()` to enforce a consistent naming scheme for functions across the package and to conform with the tidyverse style guide.
+#' 
+#' utm_to_ll takes your dataframe with UTM coordinates in separate Easting and Northing columns, and adds on an additional two columns with the converted decimalLatitude and decimalLongitude coordinates using the reference coordinate system WGS84.
+#'
+#' @details Define the name of your dataframe, the easting and northing columns within it, the UTM zone within which those coordinates are located, and the reference coordinate system (datum). UTM Northing and Easting columns must be in separate columns prior to running the function. If a datum is not defined, the function will default to "WGS84". If there are missing coordinates in your dataframe they will be preserved, however they will be moved to the end of your dataframe. Note that some parameter names are not in snake_case but instead reflect DarwinCore naming conventions.
+#'
+#' @param df - The dataframe with UTM coordinates you would like to convert. Input the name of your dataframe.
+#' @param EastingCol - The name of your Easting UTM column. Input the name in quotations, ie. "EastingCol".
+#' @param NorthingCol - The name of your Northing UTM column. Input the name in quotations, ie. "NorthingCol".
+#' @param zone - The UTM Zone. Input the zone number in quotations, ie. "17".
+#' @param datum - The datum used in the coordinate reference system of your coordinates. Input in quotations, ie. "WGS84"
+#'
+#' @return The function returns your dataframe, mutated with an additional two columns of decimal Longitude and decimal Latitude.
+#' @export
+#'
+#' @keywords internal
+#'
+#' @examples
+#' \dontrun{
+#' utm_to_ll(
+#'   df = mydataframe,
+#'   EastingCol = "EastingCoords",
+#'   NorthingCol = "NorthingCoords",
+#'   zone = "17",
+#'   datum = "WGS84"
+#' )
+#' }
+utm_to_ll <- function(df, EastingCol, NorthingCol, zone, datum = "WGS84") {
+  
+  lifecycle::deprecate_warn("0.1.0.3", "utm_to_ll()", "convert_utm_to_ll()")
+  
+  Base <- as.data.frame(df)
+  Base <- dplyr::rename(Base, "b" = EastingCol, "a" = NorthingCol)
+  Mid <- Base[!is.na(Base$"b" & Base$"a"), ]
+  Mid2 <- Base[is.na(Base$"b" & Base$"a"), ]
+  Final <- dplyr::select(Mid, "b", "a")
+  Final[1:2] <- lapply(Final[1:2], FUN = function(z) {
+    as.numeric(z)
+  })
+  
+  Final <- cbind(Final$b, Final$a)
+  
+  v <- terra::vect(Final, crs = paste0(
+    "+proj=utm +zone=",
+    zone,
+    " +datum=",
+    datum, "
+                                       +units=m"
+  ))
+  
+  converted <- terra::project(v, "+proj=longlat +datum=WGS84")
+  
+  lonlat <- terra::geom(converted)[, c("x", "y")]
+  
+  df <- cbind(Mid, lonlat)
+  df <- plyr::rbind.fill(df, Mid2)
+  df <- dplyr::rename(df,
+                      EastingCol = "b", NorthingCol = "a",
+                      "decimalLongitude" = x, "decimalLatitude" = y
   )
   return(df)
 }